Add class 'canvas' to base container in index.html
While writing the code for the function 'createGrid',
a problem came up when trying to grow the rows of
squares evenly by setting 'flex: 1;' on the 'rowOfSquares'
element. Because the CSS had a rule that set the height
of any element that was a div to 660px, it also set
the height of 'rowOfSquares,' a div, to 660px and much
larger than the base container itself. Adding a class
to the base container element named 'canvas' and changing
the CSS rule to select elements with this class to set
the height of the container fixes this problem.