Change canvas container to use min-width/min-height
The height of the canvas shrinks when the radios
are generated such that the canvas is not perfectly
square anymore. Making the initial width and height
smaller fixes this, but setting min-width/min-height
ensures the canvas will never shrink again; instead,
it will make space, even if it goes out of the viewport.