Replace radios with buttons to link with fill/erase
Every time the first canvas grid is made, none of
the radios are ever checked. This was expected, as
the user should not be able to fill or erase anything
at the start of the program. However, the radios are
not able to be deselected when a new subsequent canvas
grid is made. This causes confusion for the user, who
may believe that they can continue filling because
the radio is checked, but in reality can not because
the grid squares have been reset. Using buttons, which
do not show any indication that any choice has been
selected, fixes this problem.