Refactor fill feature to use style.cssText
The previous fill feature used classes to 'fill' the
color of the squares every time the cursor hovered
over them where every time the user hovered, an event
would be dispatched where a class named 'filled' would
be added to the event target's classList. However,
if this same method of adding classes to the event
target is to be used with the random fill feature,
the implementation would be much harder because of
the challenges of adding rules to classes dynamically
with JavaScript. Changing the background color using
style.cssText on the event target easily fixes the
problem.